+++ app/src/API/Interfaces/ForkedCommandExecutor.php
@@ -4,11 +4,11 @@ namespace Db3v4l\API\Interfaces;
 
 use Symfony\Component\Process\Process;
 
-interface ForkedSqlExecutor
+interface ForkedCommandExecutor
 {
     /**
      * @param string $sql
      * @return Process
      */
-    public function getProcess($sql);
+    public function getExecuteCommandProcess($sql);
 }