+++ app/src/Service/SqlExecutor/Forked/NativeClient.php
@@ -2,17 +2,37 @@
 
 namespace Db3v4l\Service\SqlExecutor\Forked;
 
-use Db3v4l\API\Interfaces\ForkedSqlExecutor;
+use Db3v4l\API\Interfaces\ForkedCommandExecutor;
+use Db3v4l\API\Interfaces\ForkedFileExecutor;
 use Db3v4l\Util\Process;
 
-class NativeClient extends ForkedExecutor implements ForkedSqlExecutor
+class NativeClient extends ForkedExecutor implements ForkedCommandExecutor, ForkedFileExecutor
 {
     /**
      * @param string $sql
      * @return Process
+     */
+    public function getExecuteCommandProcess($sql)
+    {
+        return $this->getProcess($sql);
+    }
+
+    /**
+     * @param string $filename
+     * @return Process
+     */
+    public function getExecuteFileProcess($filename)
+    {
+        return $this->getProcess($filename, true);
+    }
+
+    /**
+     * @param string $sqlOrFilename
+     * @param bool $isFile
+     * @return Process
      * @todo allow to inject location of db clients via setter/constructor
      */
-    public function getProcess($sql)
+    public function getProcess($sqlOrFilename, $isFile = false)
     {
         $clientType = $this->getDbClientFromDriver($this->databaseConfiguration['driver']);
 
@@ -25,9 +45,11 @@ class NativeClient extends ForkedExecutor implements ForkedSqlExecutor
                     '--user=' . $this->databaseConfiguration['user'],
                     '-p' . $this->databaseConfiguration['password'],
                     '--binary-mode', // 'It also disables all mysql commands except charset and delimiter in non-interactive mode (for input piped to mysql or loaded using the source command)'
-                    '--execute=' . $sql,
                     // $dbname
                 ];
+                if (!$isFile) {
+                    $options[] = '--execute=' . $sqlOrFilename;
+                }
                 $env = [
                     // problematic when wrapping the process in a call to `time`...
                     //'MYSQL_PWD' => $this->databaseConfiguration['password'],
@@ -41,9 +63,11 @@ class NativeClient extends ForkedExecutor implements ForkedSqlExecutor
                     //'--host=' . $this->databaseConfiguration['host'],
                     //'--port=' . $this->databaseConfiguration['port'] ?? '5432',
                     //'--username=' . $this->databaseConfiguration['user'],
-                    '--command=' . $sql,
                     //'--dbname=' . $dbname
                 ];
+                if (!$isFile) {
+                    $options[] = '--command=' . $sqlOrFilename;
+                }
                 $env = [
                     // problematic when wrapping the process in a call to `time`...
                     //'PGPASSWORD' => $this->databaseConfiguration['password'],
@@ -53,7 +77,14 @@ class NativeClient extends ForkedExecutor implements ForkedSqlExecutor
                 throw new \OutOfBoundsException("Unsupported db client '$clientType'");
         }
 
-        return new Process($this->buildCommandLine($command, $options), null, $env);
+        $commandLine = $this->buildCommandLine($command, $options);
+
+        /// @todo for psql this is probably better done via --file
+        if ($isFile) {
+            $commandLine .= ' < ' . escapeshellarg($sqlOrFilename);
+        }
+
+        return new Process($commandLine, null, $env);
     }
 
     /**